+++ src/NotificationBadge/index.js
@@ -2,9 +2,8 @@
 import * as React from "react";
 import styled from "styled-components";
 
-import Badge, { StyledBadge } from "../Badge";
+import BadgePrimitive, { StyledBadge } from "../primitives/BadgePrimitive";
 import defaultTheme from "../defaultTheme";
-import { TYPE_OPTIONS } from "../Badge/consts";
 
 import type { Props } from "./index";
 
@@ -20,13 +19,13 @@ StyledNotificationBadge.defaultProps = {
 };
 
 const NotificationBadge = (props: Props) => {
-  const { type = TYPE_OPTIONS.NEUTRAL, children, icon, ariaLabel, dataTest } = props;
+  const { type, children, icon, ariaLabel, dataTest } = props;
 
   return (
     <StyledNotificationBadge>
-      <Badge type={type} dataTest={dataTest} icon={icon} ariaLabel={ariaLabel}>
+      <BadgePrimitive type={type} dataTest={dataTest} icon={icon} ariaLabel={ariaLabel}>
         {!icon && children}
-      </Badge>
+      </BadgePrimitive>
     </StyledNotificationBadge>
   );
 };
