+++ src/PricingTable/PricingTableItem/index.js
@@ -6,7 +6,8 @@ import defaultTheme from "../../defaultTheme";
 import Stack from "../../Stack";
 import Text from "../../Text";
 import Radio from "../../Radio";
-import Badge, { StyledBadge } from "../../Badge";
+import Badge from "../../Badge";
+import { StyledBadge } from "../../primitives/BadgePrimitive";
 import media from "../../utils/mediaQuery";
 import type { Props } from "./index.js.flow";
 import STATES from "./consts";
