+++ src/primitives/BadgePrimitive/__tests__/index.test.js
@@ -3,8 +3,8 @@
 import * as React from "react";
 import { shallow } from "enzyme";
 
-import Badge from "../index";
-import Sightseeing from "../../icons/Sightseeing";
+import BadgePrimitive from "../index";
+import Sightseeing from "../../../icons/Sightseeing";
 
 describe("Badge", () => {
   const content = "badge";
@@ -12,14 +12,25 @@ describe("Badge", () => {
   const dataTest = "test";
   const icon = <Sightseeing />;
   const ariaLabel = content;
+  const background = "red";
+  const foregroundColor = "blue";
 
   const component = shallow(
-    <Badge type={type} icon={icon} dataTest={dataTest} ariaLabel={ariaLabel}>
+    <BadgePrimitive
+      type={type}
+      icon={icon}
+      background={background}
+      foregroundColor={foregroundColor}
+      dataTest={dataTest}
+      ariaLabel={ariaLabel}
+    >
       {content}
-    </Badge>,
+    </BadgePrimitive>,
   );
   it("should have passed props", () => {
     expect(component.prop("type")).toBe(type);
+    expect(component.prop("background")).toBe(background);
+    expect(component.prop("foregroundColor")).toBe(foregroundColor);
     expect(component.render().prop("data-test")).toBe(dataTest);
     expect(component.render().prop("aria-label")).toBe(ariaLabel);
   });
