+++ src/primitives/BadgePrimitive/index.js
@@ -1,0 +1,123 @@
+// @flow
+import * as React from "react";
+import styled from "styled-components";
+
+import defaultTheme from "../../defaultTheme";
+import { TYPE_OPTIONS, TOKENS } from "./consts";
+import { rtlSpacing } from "../../utils/rtl";
+
+import type { Props } from "./index";
+
+const getTypeToken = name => ({ theme, type }) => {
+  const tokens = {
+    [TOKENS.background]: {
+      [TYPE_OPTIONS.NEUTRAL]: theme.orbit.backgroundBadgeNeutral,
+      [TYPE_OPTIONS.INFO]: theme.orbit.backgroundBadgeInfo,
+      [TYPE_OPTIONS.SUCCESS]: theme.orbit.backgroundBadgeSuccess,
+      [TYPE_OPTIONS.WARNING]: theme.orbit.backgroundBadgeWarning,
+      [TYPE_OPTIONS.CRITICAL]: theme.orbit.backgroundBadgeCritical,
+      [TYPE_OPTIONS.DARK]: theme.orbit.backgroundBadgeDark,
+      [TYPE_OPTIONS.WHITE]: theme.orbit.backgroundBadgeWhite,
+      [TYPE_OPTIONS.INFO_INVERTED]: theme.orbit.paletteBlueNormal,
+      [TYPE_OPTIONS.CRITICAL_INVERTED]: theme.orbit.paletteRedNormal,
+      [TYPE_OPTIONS.SUCCESS_INVERTED]: theme.orbit.paletteGreenNormal,
+      [TYPE_OPTIONS.WARNING_INVERTED]: theme.orbit.paletteOrangeNormal,
+    },
+    [TOKENS.color]: {
+      [TYPE_OPTIONS.NEUTRAL]: theme.orbit.colorTextBadgeNeutral,
+      [TYPE_OPTIONS.INFO]: theme.orbit.colorTextBadgeInfo,
+      [TYPE_OPTIONS.SUCCESS]: theme.orbit.colorTextBadgeSuccess,
+      [TYPE_OPTIONS.WARNING]: theme.orbit.colorTextBadgeWarning,
+      [TYPE_OPTIONS.CRITICAL]: theme.orbit.colorTextBadgeCritical,
+      [TYPE_OPTIONS.DARK]: theme.orbit.colorTextBadgeDark,
+      [TYPE_OPTIONS.WHITE]: theme.orbit.colorTextBadgeWhite,
+      [TYPE_OPTIONS.INFO_INVERTED]: theme.orbit.paletteWhite,
+      [TYPE_OPTIONS.CRITICAL_INVERTED]: theme.orbit.paletteWhite,
+      [TYPE_OPTIONS.SUCCESS_INVERTED]: theme.orbit.paletteWhite,
+      [TYPE_OPTIONS.WARNING_INVERTED]: theme.orbit.paletteWhite,
+    },
+  };
+  return tokens[name][type];
+};
+
+export const StyledBadge = styled(({ className, children, dataTest, ariaLabel }) => (
+  <div className={className} data-test={dataTest} aria-label={ariaLabel}>
+    {children}
+  </div>
+))`
+  font-family: ${({ theme }) => theme.orbit.fontFamily};
+  display: inline-flex;
+  flex: 0 0 auto;
+  box-sizing: border-box;
+  justify-content: center;
+  align-items: center;
+  min-height: ${({ theme }) => theme.orbit.heightBadge};
+  line-height: 14px;
+  font-size: ${({ theme }) => theme.orbit.fontSizeTextSmall};
+  font-weight: ${({ theme }) => theme.orbit.fontWeightMedium};
+  background: ${({ background }) => background || getTypeToken(TOKENS.background)};
+  color: ${({ foregroundColor }) => foregroundColor || getTypeToken(TOKENS.color)};
+  border-radius: ${({ theme }) => theme.orbit.borderRadiusBadge};
+  padding: ${({ theme }) => theme.orbit.paddingBadge};
+`;
+
+StyledBadge.defaultProps = {
+  theme: defaultTheme,
+};
+
+const IconContainer = styled(({ className, children }) => (
+  <div className={className}>{children}</div>
+))`
+  display: flex;
+  flex-shrink: 0;
+  margin: ${({ theme, hasContent }) => hasContent && rtlSpacing(theme.orbit.marginBadgeIcon)};
+
+  svg {
+    height: ${({ theme }) => theme.orbit.widthIconSmall};
+    width: ${({ theme }) => theme.orbit.heightIconSmall};
+  }
+`;
+
+IconContainer.defaultProps = {
+  theme: defaultTheme,
+};
+
+const StyledBadgeContent = styled.div`
+  padding: 5px 0;
+  line-height: 1;
+`;
+
+StyledBadgeContent.defaultProps = {
+  theme: defaultTheme,
+};
+
+const BadgePrimitive = (props: Props) => {
+  const {
+    type = TYPE_OPTIONS.NEUTRAL,
+    icon,
+    children,
+    ariaLabel,
+    dataTest,
+    background,
+    foregroundColor,
+  } = props;
+
+  return (
+    <StyledBadge
+      type={type}
+      background={background}
+      foregroundColor={foregroundColor}
+      dataTest={dataTest}
+      ariaLabel={ariaLabel}
+    >
+      {icon && (
+        <IconContainer type={type} hasContent={!!children}>
+          {icon}
+        </IconContainer>
+      )}
+      <StyledBadgeContent>{children}</StyledBadgeContent>
+    </StyledBadge>
+  );
+};
+
+export default BadgePrimitive;
