+++ src/primitives/BadgePrimitive/index.js.flow
@@ -4,14 +4,28 @@
 */
 import { ReactComponentStyled } from "styled-components";
 
-import type { Globals } from "../common/common.js.flow";
-import type { Type } from "../primitives/BadgePrimitive";
+import type { Globals } from "../../common/common.js.flow";
+
+export type Type =
+  | "neutral"
+  | "dark"
+  | "info"
+  | "success"
+  | "warning"
+  | "critical"
+  | "white"
+  | "infoInverted"
+  | "criticalInverted"
+  | "successInverted"
+  | "warningInverted";
 
 export type Props = {|
   +children?: React$Node,
   +type?: Type,
   +icon?: React$Node,
   +ariaLabel?: string,
+  +background?: string,
+  +foregroundColor?: string,
   ...Globals,
 |};
